requirements.txt
# Core LLM framework
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# Mistral integration
langchain-mistralai>=0.1.0
mistralai>=1.0.0

# OpenRouter / OpenAI compatible client
openai>=1.30.0

# Environment variable management
python-dotenv>=1.0.1

# HTTP requests
requests>=2.31.0

# Optional but commonly used with LangChain
pydantic>=2.7.0
tiktoken>=0.7.0
streamlit>=1.20.0